Update meeting() to take start_date and end_date
These were previously called `since` and `until`, matching the other
functions, but that's misleading since the other functions allow the
records to be filtered by their modification time, but meetings can
only be filtered by the date of the meeting.

The Meeting type has an `updated` field, but the datatracker doesn't
let us filter based on that.