Increase rate limit slightly.
CircleCI runs in AWS, in an AS that IETF has seen abuse from. Accordingly,
the datatracker rate limits requests from this AS to 120 per minute. This
rate limit, of 1.5 per second, is intended to be safely below that (if the
rate limit is exceeded, the datatracker blocks access for ten minutes).